requirements.txt
absl-py==2.4.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
einops
exceptiongroup==1.3.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
grpcio==1.78.0
hypothesis==6.151.9
ImageIO==2.37.2
iniconfig==2.3.0
Jinja2==3.1.6
kiwisolver==1.4.9
lazy-loader==0.5
Markdown==3.10.2
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
packaging
pandas==2.3.3
pillow==12.1.1
pluggy==1.6.0
protobuf==7.34.0
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
scikit-image==0.25.2
scipy==1.15.3
six==1.17.0
sortedcontainers==2.4.0
sympy==1.14.0
tensorboard==2.20.0
timm
tensorboard-data-server==0.7.2
tifffile==2025.5.10
tomli==2.4.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
